requirements.txt
streamlit
pandas
st-gsheets-connection
google-generativeai
gspread